Login using axios
It's now possible for a user to log in. When the user validate the form,
a post request is sent to express server, then a 308 code is returned if
credentials are corrects or a 404 error if they are wrong. Finally vue
client handle the response with axios and redirect the user or not
according to the code returned by backend server.