UX changes on goal page
Changing the disposition of themes buttons in order to let the user scroll between them.
Also, when the user click on a theme, the other themes buttons which were opened are automatically closed.
Finally, we removed the power button of the tv for a simple led which tells if a challenge has been clickedor not.
We do not have to start the tv anymore to see challenges informations. #30